LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-window-object/close-method.window-expected.txt


FAIL window.close() queues a task to discard, but window.closed knows immediately assert_equals: expected true but got false
FAIL window.close() affects name targeting immediately assert_equals: expected true but got false
